requirements.txt
numpy
pandas
scikit-learn
matplotlib
tensorflow
tensorboard
optuna
Pillow
ipython